requirements.txt
asgiref==3.8.1
Django==5.2.2
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pydot==4.0.1
PyJWT==2.9.0
pyparsing==3.2.3
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0